Padgett's Florentine Families

florence is an adaptation of the Pajek dataset "Padgett.paj".

florence_business_ig

florence_business_nw

florence_marriage_ig

florence_marriage_nw

florence_combo_ig

florence_combo_nw

Format

Business network:

florence_business_ig

igraph

florence_business_nw

network

Marriage network:

florence_marriage_ig

igraph

florence_marriage_nw

network

Combo marriage and business:

florence_combo_ig

florence_combo_nw

Source

http://vlado.fmf.uni-lj.si/pub/networks/data/WaFa/default.htm

Details

From the Pajek documentation (originally from the UCINET IV documentation):

Breiger & Pattison (1986), in their discussion of local role analysis, use a subset of data on the social relations among Renaissance Florentine families (person aggregates) collected by John Padgett from historical documents. The two relations are business ties (PADGB - specifically, recorded financial ties such as loans, credits and joint partnerships) and marriage alliances (PADGM).

As Breiger & Pattison point out, the original data are symmetrically coded. This is acceptable perhaps for marital ties, but is unfortunate for the financial ties (which are almost certainly directed). To remedy this, the financial ties can be recoded as directed relations using some external measure of power - for instance, a measure of wealth. PADGW provides information on (1) each family's net wealth in 1427 (in thousands of lira); (2) the number of priorates (seats on the civic council) held between 1282- 1344; and (3) the total number of business or marriage ties in the total dataset of 116 families (see Breiger & Pattison (1986), p 239).

Substantively, the data include families who were locked in a struggle for political control of the city of Florence in around 1430. Two factions were dominant in this struggle: one revolved around the infamous Medicis (9), the other around the powerful Strozzis (15).

References

Breiger R. and Pattison P. (1986). Cumulated social roles: The duality of persons and their algebras. Social Networks, 8, 215-256.

Kent D. (1978). The rise of the Medici: Faction in Florence, 1426-1434. Oxford: Oxford University Press.

Examples

florence_business_ig

#> IGRAPH 1e35107 DN-- 16 30 -- 
#> + attr: name (v/c), wealth (v/n), priors (v/n), edge_type (e/c)
#> + edges from 1e35107 (vertex names):
#>  [1] Barbadori   ->Castellani   Barbadori   ->Ginori      
#>  [3] Barbadori   ->Medici       Barbadori   ->Peruzzi     
#>  [5] Bischeri    ->Guadagni     Bischeri    ->Lamberteschi
#>  [7] Bischeri    ->Peruzzi      Castellani  ->Barbadori   
#>  [9] Castellani  ->Lamberteschi Castellani  ->Peruzzi     
#> [11] Ginori      ->Barbadori    Ginori      ->Medici      
#> [13] Guadagni    ->Bischeri     Guadagni    ->Lamberteschi
#> [15] Lamberteschi->Bischeri     Lamberteschi->Castellani  
#> + ... omitted several edges

florence_business_nw

#>  Network attributes:
#>   vertices = 16 
#>   directed = TRUE 
#>   hyper = FALSE 
#>   loops = FALSE 
#>   multiple = FALSE 
#>   bipartite = FALSE 
#>   total edges= 30 
#>     missing edges= 0 
#>     non-missing edges= 30 
#> 
#>  Vertex attribute names: 
#>     priors vertex.names wealth 
#> 
#>  Edge attribute names: 
#>     edge_type 

florence_marriage_ig

#> IGRAPH 1e349c1 DN-- 16 40 -- 
#> + attr: name (v/c), wealth (v/n), priors (v/n), edge_type (e/c)
#> + edges from 1e349c1 (vertex names):
#>  [1] Acciaiuoli  ->Medici       Albizzi     ->Ginori      
#>  [3] Albizzi     ->Guadagni     Albizzi     ->Medici      
#>  [5] Barbadori   ->Castellani   Barbadori   ->Medici      
#>  [7] Bischeri    ->Guadagni     Bischeri    ->Peruzzi     
#>  [9] Bischeri    ->Strozzi      Castellani  ->Barbadori   
#> [11] Castellani  ->Peruzzi      Castellani  ->Strozzi     
#> [13] Ginori      ->Albizzi      Guadagni    ->Albizzi     
#> [15] Guadagni    ->Bischeri     Guadagni    ->Lamberteschi
#> + ... omitted several edges

florence_marriage_nw

#>  Network attributes:
#>   vertices = 16 
#>   directed = TRUE 
#>   hyper = FALSE 
#>   loops = FALSE 
#>   multiple = FALSE 
#>   bipartite = FALSE 
#>   total edges= 40 
#>     missing edges= 0 
#>     non-missing edges= 40 
#> 
#>  Vertex attribute names: 
#>     priors vertex.names wealth 
#> 
#>  Edge attribute names: 
#>     edge_type 

florence_combo_ig

#> IGRAPH 1e355eb DN-- 16 70 -- 
#> + attr: name (v/c), wealth (v/n), priors (v/n), edge_type (e/c)
#> + edges from 1e355eb (vertex names):
#>  [1] Barbadori   ->Castellani   Barbadori   ->Ginori      
#>  [3] Barbadori   ->Medici       Barbadori   ->Peruzzi     
#>  [5] Bischeri    ->Guadagni     Bischeri    ->Lamberteschi
#>  [7] Bischeri    ->Peruzzi      Castellani  ->Barbadori   
#>  [9] Castellani  ->Lamberteschi Castellani  ->Peruzzi     
#> [11] Ginori      ->Barbadori    Ginori      ->Medici      
#> [13] Guadagni    ->Bischeri     Guadagni    ->Lamberteschi
#> [15] Lamberteschi->Bischeri     Lamberteschi->Castellani  
#> + ... omitted several edges

florence_combo_nw

#>  Network attributes:
#>   vertices = 16 
#>   directed = TRUE 
#>   hyper = FALSE 
#>   loops = FALSE 
#>   multiple = TRUE 
#>   bipartite = FALSE 
#>   total edges= 70 
#>     missing edges= 0 
#>     non-missing edges= 70 
#> 
#>  Vertex attribute names: 
#>     priors vertex.names wealth 
#> 
#>  Edge attribute names: 
#>     edge_type