Koschade's Jemaah Islamiyah (2002 Bali Bombings)

jemmah_islamiyah is an adaptation of the UCINET dataset "Jemaah Islamiyah Koschade".

jemaah_islamiyah_ig

jemaah_islamiyah_nw

Format

jemaah_islamiyah_ig: igraph

jemaah_islamiyah_nw: network

Source

https://sites.google.com/site/ucinetsoftware/datasets/covert-networks/jemaahislamiyahkoschade

Details

From the UCINET documentation:

Jemaah Islamiyah cell that was responsible for the Bali bombings in 2002 – i.e. should tally with the other Bali bombing dataset from JJATT. The recording of the interaction of the cell began following the meeting in the Hotel Harem in Denpasar on October 6, when the group was considered to go ‘operationally covert’, and concluded when the majority of the group had left Bali before the implementation of the operation on October 11, 2002.

References

Koschade, Stuart (2006) A Social Network Analysis of Jemaah Islamiyah: The Applications to Counter-Terrorism and Intelligence. Studies in Conflict and Terrorism Vol. 29(6):pp. 559-575

International Crisis Group. 2003 ‘Jemaah Islamiyah in South East Asia: Damaged but Still Dangerous’

Examples

jemaah_islamiyah_ig

#> IGRAPH b61f3bc UNW- 17 63 -- 
#> + attr: name (v/c), weight (e/n)
#> + edges from b61f3bc (vertex names):
#>  [1] MUKLAS --AMROZI   MUKLAS --IMRON    MUKLAS --SAMUDRA  MUKLAS --DULMATIN
#>  [5] MUKLAS --IDRIS    MUKLAS --AZAHARI  MUKLAS --GHONI    MUKLAS --PATEK   
#>  [9] MUKLAS --SARIJO   AMROZI --SAMUDRA  AMROZI --IDRIS    AMROZI --MUBAROK 
#> [13] IMRON  --SAMUDRA  IMRON  --DULMATIN IMRON  --IDRIS    IMRON  --AZAHARI 
#> [17] IMRON  --GHONI    IMRON  --PATEK    IMRON  --FERI     IMRON  --SARIJO  
#> [21] SAMUDRA--DULMATIN SAMUDRA--IDRIS    SAMUDRA--MUBAROK  SAMUDRA--AZAHARI 
#> [25] SAMUDRA--GHONI    SAMUDRA--ARNASAN  SAMUDRA--RAUF     SAMUDRA--OCTAVIA 
#> [29] SAMUDRA--HIDAYAT  SAMUDRA--JUNAEDI  SAMUDRA--PATEK    SAMUDRA--SARIJO  
#> + ... omitted several edges

jemaah_islamiyah_nw

#>  Network attributes:
#>   vertices = 17 
#>   directed = FALSE 
#>   hyper = FALSE 
#>   loops = FALSE 
#>   multiple = FALSE 
#>   bipartite = FALSE 
#>   total edges= 63 
#>     missing edges= 0 
#>     non-missing edges= 63 
#> 
#>  Vertex attribute names: 
#>     vertex.names 
#> 
#>  Edge attribute names: 
#>     weight